tools/beca_link/requirements.txt
pyserial==3.5
mido==1.3.3
python-rtmidi==1.5.8
